Make count int64 function

+++ sql/expression/function/aggregation/count_test.go
@@ -21,20 +21,20 @@ func TestCount_Eval_1(t *testing.T) {
 
 	c := NewCount(expression.NewLiteral(1, sql.Int32))
 	b := c.NewBuffer()
-	require.Equal(int32(0), eval(t, c, b))
+	require.Equal(int64(0), eval(t, c, b))
 
 	require.NoError(c.Update(ctx, b, nil))
 	require.NoError(c.Update(ctx, b, sql.NewRow("foo")))
 	require.NoError(c.Update(ctx, b, sql.NewRow(1)))
 	require.NoError(c.Update(ctx, b, sql.NewRow(nil)))
 	require.NoError(c.Update(ctx, b, sql.NewRow(1, 2, 3)))
-	require.Equal(int32(5), eval(t, c, b))
+	require.Equal(int64(5), eval(t, c, b))
 
 	b2 := c.NewBuffer()
 	require.NoError(c.Update(ctx, b2, nil))
 	require.NoError(c.Update(ctx, b2, sql.NewRow("foo")))
 	require.NoError(c.Merge(ctx, b, b2))
-	require.Equal(int32(7), eval(t, c, b))
+	require.Equal(int64(7), eval(t, c, b))
 }
 
 func TestCount_Eval_Star(t *testing.T) {
@@ -43,20 +43,20 @@ func TestCount_Eval_Star(t *testing.T) {
 
 	c := NewCount(expression.NewStar())
 	b := c.NewBuffer()
-	require.Equal(int32(0), eval(t, c, b))
+	require.Equal(int64(0), eval(t, c, b))
 
 	c.Update(ctx, b, nil)
 	c.Update(ctx, b, sql.NewRow("foo"))
 	c.Update(ctx, b, sql.NewRow(1))
 	c.Update(ctx, b, sql.NewRow(nil))
 	c.Update(ctx, b, sql.NewRow(1, 2, 3))
-	require.Equal(int32(5), eval(t, c, b))
+	require.Equal(int64(5), eval(t, c, b))
 
 	b2 := c.NewBuffer()
 	c.Update(ctx, b2, sql.NewRow())
 	c.Update(ctx, b2, sql.NewRow("foo"))
 	c.Merge(ctx, b, b2)
-	require.Equal(int32(7), eval(t, c, b))
+	require.Equal(int64(7), eval(t, c, b))
 }
 
 func TestCount_Eval_String(t *testing.T) {
@@ -65,11 +65,11 @@ func TestCount_Eval_String(t *testing.T) {
 
 	c := NewCount(expression.NewGetField(0, sql.Text, "", true))
 	b := c.NewBuffer()
-	require.Equal(int32(0), eval(t, c, b))
+	require.Equal(int64(0), eval(t, c, b))
 
 	c.Update(ctx, b, sql.NewRow("foo"))
-	require.Equal(int32(1), eval(t, c, b))
+	require.Equal(int64(1), eval(t, c, b))
 
 	c.Update(ctx, b, sql.NewRow(nil))
-	require.Equal(int32(1), eval(t, c, b))
+	require.Equal(int64(1), eval(t, c, b))
 }
